pyproject.toml
[project]
name = "skiller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.4.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.14.3",
    "chonkie[semantic]>=1.5.2",
    "firecrawl-py>=4.13.0",
    "google-genai>=1.59.0",
    "google-generativeai>=0.8.6",
    "lancedb>=0.26.1",
    "langwatch>=0.9.0",
    "mistralai>=1.10.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pylance>=1.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "supermemory>=3.18.0",
    "tantivy>=0.25.1",
    "tweepy>=4.16.0",
    "openai>=2.15.0",
    "apify-client>=1.6.0",
    "scrapegraph-py>=1.0.0",
    "scrapebadger>=0.1.7",
    "uvicorn>=0.40.0",
    "fastapi>=0.128.0",
    "sqlalchemy>=2.0.46",
    "pyjwt>=2.10.1",
    "structlog>=25.5.0",
    "psutil>=7.2.1",
    "cli2>=5.2.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = ["prompts/*.yaml", "prompts/.materialized/*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
skiller = "app.main:main"
